[Chore] #99 - 사용하지 않는 코드/파일 삭제
