--- ToDwoong/ToDwoong/Global/Protocols/AddTodoGroupSelectControllerDelegate.swift
@@ -1,10 +1,0 @@
-//
-//  AddTodoGroupSelectControllerDelegate.swift
-//  ToDwoong
-//
-//  Created by mirae on 3/5/24.
-//
-
-protocol AddTodoGroupSelectControllerDelegate: AnyObject {
-    func groupSelectController(_ controller: AddTodoGroupSelectController, didSelectGroup category: Category)
-}
