--- ToDwoong/ToDwoong/Global/Protocols/AddTodoLocationPickerDelegate.swift
@@ -1,10 +1,0 @@
-//
-//  AddTodoLocationPickerDelegate.swift
-//  ToDwoong
-//
-//  Created by mirae on 3/5/24.
-//
-
-protocol AddTodoLocationPickerDelegate: AnyObject {
-    func didPickLocation(_ address: String)
-}
