--- ToDwoong/ToDwoong/Global/Protocols/AddTodoLocationPickerViewDelegate.swift
@@ -1,10 +1,0 @@
-//
-//  AddTodoLocationPickerViewDelegate.swift
-//  ToDwoong
-//
-//  Created by mirae on 3/5/24.
-//
-
-protocol AddTodoLocationPickerViewDelegate: AnyObject {
-    func didTapConfirmAddress(_ address: String)
-}
