--- ToDwoong/ToDwoong/Global/Protocols/TitleCollectionViewCellDelegate.swift
@@ -1,10 +1,0 @@
-//
-//  TitleCollectionViewCellDelegate.swift
-//  ToDwoong
-//
-//  Created by mirae on 3/5/24.
-//
-
-protocol TitleCollectionViewCellDelegate: AnyObject {
-    func titleCellDidEndEditing(_ text: String?)
-}
