+++ ToDwoong/ToDwoong/Presentation/TabBar/TabBarController.swift
@@ -9,20 +9,6 @@ import UIKit
 
 import TodwoongDesign
 
-class FirstViewController: UIViewController {
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        view.backgroundColor = TDStyle.color.bgBlue
-    }
-}
-
-class SecondViewController: UIViewController {
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        view.backgroundColor = TDStyle.color.bgRed
-    }
-}
-
 class TabBarController: UITabBarController {
     override func viewDidLoad() {
         super.viewDidLoad()
